Move more legacy in-memory trace code to trace-legacy.{c++,h}.
