+++ src/workerd/io/trace-common.h
@@ -4,6 +4,7 @@
 #include <workerd/io/worker-interface.capnp.h>
 #include <workerd/jsg/memory.h>
 
+#include <kj/map.h>
 #include <kj/one-of.h>
 #include <kj/string.h>
 #include <kj/time.h>
@@ -241,6 +242,40 @@ struct Exception final {
   kj::Maybe<kj::String> stack;
 
   void copyTo(rpc::Trace::Exception::Builder builder);
+};
+
+// Represents a trace span. `Span` objects are delivered to `SpanObserver`s for recording. To
+// create a `Span`, use a `SpanBuilder`.
+struct Span {
+  using TagValue = kj::OneOf<bool, int64_t, double, kj::String>;
+  // TODO(someday): Support binary bytes, too.
+  using TagMap = kj::HashMap<kj::ConstString, TagValue>;
+  using Tag = TagMap::Entry;
+
+  struct Log {
+    kj::Date timestamp;
+    Tag tag;
+  };
+
+  kj::ConstString operationName;
+  kj::Date startTime;
+  kj::Date endTime;
+  TagMap tags;
+  kj::Vector<Log> logs;
+
+  // We set an arbitrary (-ish) cap on log messages for safety. If we drop logs because of this,
+  // we report how many in a final "dropped_logs" log.
+  //
+  // At the risk of being too clever, I chose a limit that is one below a power of two so that
+  // we'll typically have space for one last element available for the "dropped_logs" log without
+  // needing to grow the vector.
+  static constexpr auto MAX_LOGS = 1023;
+  uint droppedLogs = 0;
+
+  explicit Span(kj::ConstString operationName, kj::Date startTime)
+      : operationName(kj::mv(operationName)),
+        startTime(startTime),
+        endTime(startTime) {}
 };
 
 }  // namespace trace
